refactor: image getters in Product model now return empty list instead of None. Add getters for image urls list.
If the image list is empty, the getters now return an empty list instead of None. This makes more sense, since they'll be mainly used in forloops in templates and this makes it more transparent, since no if checks to make sure that the list is not None are needed.